fix： 1. 剥离shell函数到util.cmd包下，作为公共模块使用； 2. 调用shell方法获取es_gateway所在的主目录，自动配置gateway_conf
