+++ es_gateway/utils/nginx_signals.lua
@@ -11,7 +11,10 @@ local gw_conf = require "es_gateway.gateway_conf"
 local kill = require "es_gateway.utils.kill"
 local logger = require "es_gateway.utils.logger"
 local verison = require "es_gateway.utils.version"
+local sh =  require "es_gateway.utils.cmd.shell"
+local shell = sh.shell
 local fmt = string.format
+
 
 local nginx_bin_name = "nginx"
 local nginx_helper_shell_name = "default_nginx.sh"
@@ -29,37 +32,37 @@ local nginx_compatible = unpack(meta._DEPENDENCIES.openresty)
 --   @param s: stirng
 -- @returns
 --   @param o: cmd output  
-local function trim_and_delete_LF(s)
-    return string.gsub(s, "^%s*(.-)%s*[\n]*$", "%1")
-end
+--local function trim_and_delete_LF(s)
+--    return string.gsub(s, "^%s*(.-)%s*[\n]*$", "%1")
+--end
 
 -- @func shell: Perform a shell command and return its output
 --   @param c: command
 -- @returns
 --   @param o: output, or nil if error
 -- NOTES: out_file_name out path is: /home/sm01/openresty-1.11.2/lua.output
-local function shell (c)
-    if not c then 
-        logger.warn("shell cmd is nil")
-        return nil  
-    end
-    local out_file_name = "lua.output"
-    local ok = os.execute (fmt("%s > %s 2>&1", c, out_file_name))
-    local fd
-    local o
-    if ok then
-        fd = assert(io.open(out_file_name))
-        o = fd:read ("*a")
-        fd:close ()
-    end
-
-    -- delete out_file_name
-    os.remove(out_file_name)
-    if not o then
-        return nil
-    end
-    return trim_and_delete_LF(o)
-end
+--local function shell (c)
+--    if not c then
+--        logger.warn("shell cmd is nil")
+--        return nil
+--    end
+--    local out_file_name = "lua.output"
+--    local ok = os.execute (fmt("%s > %s 2>&1", c, out_file_name))
+--    local fd
+--    local o
+--    if ok then
+--        fd = assert(io.open(out_file_name))
+--        o = fd:read ("*a")
+--        fd:close ()
+--    end
+--
+--     delete out_file_name
+--    os.remove(out_file_name)
+--    if not o then
+--        return nil
+--    end
+--    return trim_and_delete_LF(o)
+--end
 
 --
 --  @func is_openresty: Test if this openresty verison is compatiable with meta._DEPENDENCIES.openresty
